IntensityScans

IntensityScans.IntensityScan

IntensityScans.scan_x

IntensityScans.scan_xy

IntensityScans.scan_xyz

IntensityScans.scan_y

IntensityScans.scan_z

IntensityScans.IntensityScan — Type

Input: xyz: The handle of desired xyz stage scope: Handle of desired scope channel: This input decides which channel the scope will read from. sample_size: The amount of samples desired to be taken.

Ex.

lts = initialize(ThorlabsLTS150)
scope = initialize(Scope350MHz)
# Sample size 100, read from scope on channel 1
hydrophone = IntensityScan(lts, scope, 100, 1)

IntensityScans.scan_x — Method

scan_x(scanner, x_range, x_num_scans; kwargs..)

Move along the x axis and grab data based on a fixed increment.

Arguments

scanner::IntensityScan: the structure for all controller connections.

x_range: An array or tuple that contains the start and end point of the desired range in the x-direction. (Must be 1 or 2 elements long)

x_num_scans::Int: The number of points/scans the user desires to record along the x-axis.

Keywords

verbose::Bool: Optional, defaults to: true

Returns

Scan1D

Example

xyz_handle = initialize(ThorlabsLTS150)
scope_handle = initialize(Scope350MHz)
# Sample size 100, read from scope on channel 1
scanner_handle = IntensityScan(xyz_handle, scope_handle, 100, 1)

scan_range_x = [0, 2]
num_points_x = 5

scan_info_x = scan_x(scanner_handle, scan_range_x, num_points_x)

IntensityScans.scan_xy — Method

scan_xy(scanner, x_axis_range, x_axis_num_scans, y_axis_range, y_axis_num_scans; kwargs..)

This function grabs data along the xy-plane. It holds the y-position constant and moves along the x axis until it reaches the last x point and then changes the y-position and repeats. There are five mandatory inputs and one optional input that is used only when this function is called by other functions and the printed messages should be surpressed. The user usually doesn't have to use the variable called.

Arguments

scanner: the structure for all controller connections.

x_axis_range::Array: An array that contains the start and end point of the desired range in the x-direction. (Must be 1 or 2 elements long)

x_axis_num_scans::Int: The number of points/scans the user desires to record along the x-axis.

y_axis_range::Array: Contains the start and end point of the desired range in the y-direction. (Must be 1 or 2 elements long)

y_axis_num_scans::Int: The number of points/scans the user desires to record along the y-axis.

Keywords

verbose::Bool: Optional, defaults to: true

Returns

Scan2D

Example

scanner = IntensityScan(initialize(lts150), initialize(Scope350MHz), 1, 100)
scan_range_x = [0 2]; scan_range_y = [0 2];
num_points_x = 5; num_points_y =10;

wave_info_xz = scan_xy(
    scanner, scan_range_x, num_points_x, scan_range_y, num_points_y
)

IntensityScans.scan_xyz — Method

TODO: Rewrite function waveinfovol = kyhydroscanvolume(settings, xrange, xnum, yrange, ynum, zrange, znum, called) waveinfovol = kyhydroscanvolume(settings, xrange, xnum, yrange, ynum, zrange, znum, called)

This function grabs data along the xyz-axis. It holds the z-position constant and calls the xy_scan function. This function scans along the x axis until it reaches the last x point and then moves along the y axis and repeats the process. After completing the xy scan, the z-position is changed and then the process is repeated. There are five mantadory inputs and one optional input that is used only when this function is called by other functions and the printed messages should be surpressed. The user usually doesn't have to use the variable called.

Input: settings: The structure for all controller connections. xrange: A 1x2 or 1x1 array that contains the start and end point of the desired range in the x-direction. xnum: The number of points the user desires to record along the x-axis. yrange: A 1x2 or 1x1 array that contains the start and end point of the desired range in the y-direction. ynum: The number of points the user desires to record along the y-axis. zrange: A 1x2 or 1x1 array that contains the start and end point of the desired range in the z-direction. znum: The number of points the user desires to record along the z-axis.

Output: waveinfoxy: A structure that contains the data of the waveforms and the coordinates travelled. The waveforms are stored under waveinfovol.waveform and the coordinates travelled are stored under waveinfo_vol.coordinates. The coordinates are the (x, y, z) coordinates along the path that the function took.

Example: scanrangex = [0 2]; scanrangey = [0 2]; scanrangez = [0 2]; numpointsx = 5; numpointsy = 10; numpointsz = 10; waveinfovol = kyhydroscanvolume(settings, scanrangex, numpointsx, scanrangey, numpointsy, scanrangez, numpoints_z)

IntensityScans.scan_y — Method

scan_y(scanner, y_axis_range, y_axis_num_scans; kwargs..)

This function moves along the y axis and grabs data based on a fixed increment passed by the user.

Arguments

scanner: the structure for all controller connections.

y_axis_range::Array: Contains the start and end point of the desired range in the y-direction. (Must be 1 or 2 elements long)

y_axis_num_scans::Int: The number of points/scans the user desires to record along the y-axis.

Keywords

verbose::Bool: Optional, defaults to: true

Returns

Scan1D

Example

xyz_handle = initialize(ThorlabsLTS150)
scope_handle = initialize(Scope350MHz)
# Sample size 100, read from scope on channel 1
scan_handle = IntensityScan(xyz_handle, scope_handle, 100, 1)

scan_range_y = [0 2]
num_points_y =10

scan_info_y = scan_y(scan_handle, scan_range_y, num_points_y)

IntensityScans.scan_z — Method

scan_z(scan_handle, z_scan_range, num_scans)

This function moves along the z axis and grabs data based on a fixed increment passed by the user.

Arguments

scanner: the structure for all controller connections.

z_axis_range::Array: Contains the start and end point of the desired range in the z-direction. (Must be 1 or 2 elements long)

z_axis_num_scans::Int: The number of points/scans the user desires to record along the z-axis.

Keywords

verbose: Default true. When false the printed messages should be suppressed (normally used when this function is called by another function)

Returns

Scan1D

Example

xyz_handle = initialize(ThorlabsLTS150)
scope_handle = initialize(Scope350MHz)
# Sample size 100, read from scope on channel 1
scan_handle = IntensityScan(xyz_handle, scope_handle, 100, 1)

scan_range = [0, 1]
num_scans = 5
scan_info_z = scan_z(scan_handle, scan_range, num_scans)